Revert "Merge pull request #165 from launchdarkly/eb/ch59586/integrations-package"
This reverts commit 9cd52288791afd25bc7ff0dfccc4fff67a87fd4d, reversing
changes made to c273a05f3241b7e482d4fdb78cb9d684529b1fc4.
