+++ src/main/java/com/launchdarkly/client/RedisFeatureStoreBuilder.java
@@ -1,23 +1,25 @@
 package com.launchdarkly.client;
 
-import com.launchdarkly.client.integrations.Redis;
-import com.launchdarkly.client.integrations.RedisDataStoreBuilder;
+import redis.clients.jedis.JedisPoolConfig;
+import redis.clients.jedis.Protocol;
 
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.concurrent.TimeUnit;
 
-import redis.clients.jedis.JedisPoolConfig;
-
 /**
- * Deprecated builder class for the Redis-based persistent data store.
- * <p>
- * The replacement for this class is {@link com.launchdarkly.client.integrations.RedisDataStoreBuilder}.
- * This class is retained for backward compatibility and will be removed in a future version. 
- * 
- * @deprecated Use {@link com.launchdarkly.client.integrations.Redis#dataStore()}
+ * A <a href="http://en.wikipedia.org/wiki/Builder_pattern">builder</a> for configuring the Redis-based persistent feature store.
+ *
+ * Obtain an instance of this class by calling {@link Components#redisFeatureStore()} or {@link Components#redisFeatureStore(URI)}.
+ * Builder calls can be chained, for example:
+ *
+ * <pre><code>
+ * FeatureStore store = Components.redisFeatureStore()
+ *      .database(1)
+ *      .caching(FeatureStoreCacheConfig.enabled().ttlSeconds(60))
+ *      .build();
+ * </code></pre>
  */
-@Deprecated
 public final class RedisFeatureStoreBuilder implements FeatureStoreFactory {
   /**
    * The default value for the Redis URI: {@code redis://localhost:6379}
@@ -38,21 +40,25 @@ public final class RedisFeatureStoreBuilder implements FeatureStoreFactory {
    */
   public static final long DEFAULT_CACHE_TIME_SECONDS = FeatureStoreCacheConfig.DEFAULT_TIME_SECONDS;
   
-  final RedisDataStoreBuilder wrappedBuilder;
-  
-  // We have to keep track of these caching parameters separately in order to support some deprecated setters
+  final URI uri;
+  String prefix = DEFAULT_PREFIX;
+  int connectTimeout = Protocol.DEFAULT_TIMEOUT;
+  int socketTimeout = Protocol.DEFAULT_TIMEOUT;
+  Integer database = null;
+  String password = null;
+  boolean tls = false;
   FeatureStoreCacheConfig caching = FeatureStoreCacheConfig.DEFAULT;
-  boolean refreshStaleValues = false;
+  boolean refreshStaleValues = false; // this and asyncRefresh are redundant with FeatureStoreCacheConfig, but are used by deprecated setters
   boolean asyncRefresh = false;
+  JedisPoolConfig poolConfig = null;
 
-  // These constructors are called only from Components
+  // These constructors are called only from Implementations
   RedisFeatureStoreBuilder() {
-    wrappedBuilder = Redis.dataStore();
+    this.uri = DEFAULT_URI;
   }
   
   RedisFeatureStoreBuilder(URI uri) {
-    this();
-    wrappedBuilder.uri(uri);
+    this.uri = uri;
   }
   
   /**
@@ -63,10 +69,8 @@ public final class RedisFeatureStoreBuilder implements FeatureStoreFactory {
    * @deprecated Please use {@link Components#redisFeatureStore(java.net.URI)}.
    */
   public RedisFeatureStoreBuilder(URI uri, long cacheTimeSecs) {
-    this();
-    wrappedBuilder.uri(uri);
-    caching = caching.ttlSeconds(cacheTimeSecs);
-    wrappedBuilder.caching(caching);
+    this.uri = uri;
+    this.cacheTime(cacheTimeSecs, TimeUnit.SECONDS);
   }
 
   /**
@@ -80,10 +84,8 @@ public final class RedisFeatureStoreBuilder implements FeatureStoreFactory {
    * @deprecated Please use {@link Components#redisFeatureStore(java.net.URI)}.
    */
   public RedisFeatureStoreBuilder(String scheme, String host, int port, long cacheTimeSecs) throws URISyntaxException {
-    this();
-    wrappedBuilder.uri(new URI(scheme, null, host, port, null, null, null));
-    caching = caching.ttlSeconds(cacheTimeSecs);
-    wrappedBuilder.caching(caching);
+    this.uri = new URI(scheme, null, host, port, null, null, null);
+    this.cacheTime(cacheTimeSecs, TimeUnit.SECONDS);
   }
 
   /**
@@ -98,7 +100,7 @@ public final class RedisFeatureStoreBuilder implements FeatureStoreFactory {
    * @since 4.7.0
    */
   public RedisFeatureStoreBuilder database(Integer database) {
-    wrappedBuilder.database(database);
+    this.database = database;
     return this;
   }
   
@@ -114,7 +116,7 @@ public final class RedisFeatureStoreBuilder implements FeatureStoreFactory {
    * @since 4.7.0
    */
   public RedisFeatureStoreBuilder password(String password) {
-    wrappedBuilder.password(password);
+    this.password = password;
     return this;
   }
   
@@ -131,7 +133,7 @@ public final class RedisFeatureStoreBuilder implements FeatureStoreFactory {
    * @since 4.7.0
    */
   public RedisFeatureStoreBuilder tls(boolean tls) {
-    wrappedBuilder.tls(tls);
+    this.tls = tls;
     return this;
   }
   
@@ -147,7 +149,6 @@ public final class RedisFeatureStoreBuilder implements FeatureStoreFactory {
    */
   public RedisFeatureStoreBuilder caching(FeatureStoreCacheConfig caching) {
     this.caching = caching;
-    wrappedBuilder.caching(caching);
     return this;
   }
   
@@ -186,14 +187,13 @@ public final class RedisFeatureStoreBuilder implements FeatureStoreFactory {
   private void updateCachingStaleValuesPolicy() {
     // We need this logic in order to support the existing behavior of the deprecated methods above:
     // asyncRefresh is supposed to have no effect unless refreshStaleValues is true
-    if (refreshStaleValues) {
-      caching = caching.staleValuesPolicy(this.asyncRefresh ?
+    if (this.refreshStaleValues) {
+      this.caching = this.caching.staleValuesPolicy(this.asyncRefresh ?
           FeatureStoreCacheConfig.StaleValuesPolicy.REFRESH_ASYNC :
           FeatureStoreCacheConfig.StaleValuesPolicy.REFRESH);
     } else {
-      caching = caching.staleValuesPolicy(FeatureStoreCacheConfig.StaleValuesPolicy.EVICT);
+      this.caching = this.caching.staleValuesPolicy(FeatureStoreCacheConfig.StaleValuesPolicy.EVICT);
     }
-    wrappedBuilder.caching(caching);
   }
   
   /**
@@ -203,7 +203,7 @@ public final class RedisFeatureStoreBuilder implements FeatureStoreFactory {
    * @return the builder
    */
   public RedisFeatureStoreBuilder prefix(String prefix) {
-    wrappedBuilder.prefix(prefix);
+    this.prefix = prefix;
     return this;
   }
 
@@ -218,9 +218,8 @@ public final class RedisFeatureStoreBuilder implements FeatureStoreFactory {
    * @deprecated use {@link #caching(FeatureStoreCacheConfig)} and {@link FeatureStoreCacheConfig#ttl(long, TimeUnit)}.
    */
   public RedisFeatureStoreBuilder cacheTime(long cacheTime, TimeUnit timeUnit) {
-    caching = caching.ttl(cacheTime, timeUnit)
+    this.caching = this.caching.ttl(cacheTime, timeUnit)
         .staleValuesPolicy(this.caching.getStaleValuesPolicy());
-    wrappedBuilder.caching(caching);
     return this;
   }
 
@@ -231,7 +230,7 @@ public final class RedisFeatureStoreBuilder implements FeatureStoreFactory {
    * @return the builder
    */
   public RedisFeatureStoreBuilder poolConfig(JedisPoolConfig poolConfig) {
-    wrappedBuilder.poolConfig(poolConfig);
+    this.poolConfig = poolConfig;
     return this;
   }
 
@@ -244,7 +243,7 @@ public final class RedisFeatureStoreBuilder implements FeatureStoreFactory {
    * @return the builder
    */
   public RedisFeatureStoreBuilder connectTimeout(int connectTimeout, TimeUnit timeUnit) {
-    wrappedBuilder.connectTimeout(connectTimeout, timeUnit);
+    this.connectTimeout = (int) timeUnit.toMillis(connectTimeout);
     return this;
   }
 
@@ -257,7 +256,7 @@ public final class RedisFeatureStoreBuilder implements FeatureStoreFactory {
    * @return the builder
    */
   public RedisFeatureStoreBuilder socketTimeout(int socketTimeout, TimeUnit timeUnit) {
-    wrappedBuilder.socketTimeout(socketTimeout, timeUnit);
+    this.socketTimeout = (int) timeUnit.toMillis(socketTimeout);
     return this;
   }
