+++ src/main/java/com/launchdarkly/client/files/FileDataSource.java
@@ -1,34 +1,21 @@
-package com.launchdarkly.client.integrations;
+package com.launchdarkly.client.files;
 
 import com.google.common.util.concurrent.Futures;
-import com.google.gson.JsonElement;
 import com.launchdarkly.client.FeatureStore;
 import com.launchdarkly.client.UpdateProcessor;
-import com.launchdarkly.client.VersionedData;
-import com.launchdarkly.client.VersionedDataKind;
-import com.launchdarkly.client.integrations.FileDataSourceParsing.FileDataException;
-import com.launchdarkly.client.integrations.FileDataSourceParsing.FlagFactory;
-import com.launchdarkly.client.integrations.FileDataSourceParsing.FlagFileParser;
-import com.launchdarkly.client.integrations.FileDataSourceParsing.FlagFileRep;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.nio.file.FileSystem;
 import java.nio.file.FileSystems;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.WatchEvent;
 import java.nio.file.WatchKey;
 import java.nio.file.WatchService;
 import java.nio.file.Watchable;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicBoolean;
@@ -38,20 +25,20 @@ import static java.nio.file.StandardWatchEventKinds.ENTRY_DELETE;
 import static java.nio.file.StandardWatchEventKinds.ENTRY_MODIFY;
 
 /**
- * Implements taking flag data from files and putting it into the data store, at startup time and
+ * Implements taking flag data from files and putting it into the feature store, at startup time and
  * optionally whenever files change.
  */
-final class FileDataSourceImpl implements UpdateProcessor {
-  private static final Logger logger = LoggerFactory.getLogger(FileDataSourceImpl.class);
+class FileDataSource implements UpdateProcessor {
+  private static final Logger logger = LoggerFactory.getLogger(FileDataSource.class);
 
   private final FeatureStore store;
   private final DataLoader dataLoader;
   private final AtomicBoolean inited = new AtomicBoolean(false);
   private final FileWatcher fileWatcher;
   
-  FileDataSourceImpl(FeatureStore store, List<Path> sources, boolean autoUpdate) {
+  FileDataSource(FeatureStore store, DataLoader dataLoader, boolean autoUpdate) {
     this.store = store;
-    this.dataLoader = new DataLoader(sources);
+    this.dataLoader = dataLoader;
 
     FileWatcher fw = null;
     if (autoUpdate) {
@@ -78,7 +65,7 @@ final class FileDataSourceImpl implements UpdateProcessor {
     if (fileWatcher != null) {
       fileWatcher.start(new Runnable() {
         public void run() {
-          FileDataSourceImpl.this.reload();
+          FileDataSource.this.reload();
         }
       });
     }
@@ -90,7 +77,7 @@ final class FileDataSourceImpl implements UpdateProcessor {
     DataBuilder builder = new DataBuilder(); 
     try {
       dataLoader.load(builder); 
-    } catch (FileDataException e) {
+    } catch (DataLoaderException e) {
       logger.error(e.getDescription());
       return false;
     }
@@ -114,7 +101,7 @@ final class FileDataSourceImpl implements UpdateProcessor {
   /**
    * If auto-updating is enabled, this component watches for file changes on a worker thread.
    */
-  private static final class FileWatcher implements Runnable {
+  private static class FileWatcher implements Runnable {
     private final WatchService watchService;
     private final Set<Path> watchedFilePaths;
     private Runnable fileModifiedAction;
@@ -178,7 +165,7 @@ final class FileDataSourceImpl implements UpdateProcessor {
     
     public void start(Runnable fileModifiedAction) {
       this.fileModifiedAction = fileModifiedAction;
-      thread = new Thread(this, FileDataSourceImpl.class.getName());
+      thread = new Thread(this, FileDataSource.class.getName());
       thread.setDaemon(true);
       thread.start();
     }
@@ -188,77 +175,6 @@ final class FileDataSourceImpl implements UpdateProcessor {
       if (thread != null) {
         thread.interrupt();
       }
-    }
-  }
-  
-  /**
-   * Implements the loading of flag data from one or more files. Will throw an exception if any file can't
-   * be read or parsed, or if any flag or segment keys are duplicates.
-   */
-  static final class DataLoader {
-    private final List<Path> files;
-
-    public DataLoader(List<Path> files) {
-      this.files = new ArrayList<Path>(files);
-    }
-    
-    public Iterable<Path> getFiles() {
-      return files;
-    }
-    
-    public void load(DataBuilder builder) throws FileDataException
-    {
-      for (Path p: files) {
-        try {
-          byte[] data = Files.readAllBytes(p);
-          FlagFileParser parser = FlagFileParser.selectForContent(data);
-          FlagFileRep fileContents = parser.parse(new ByteArrayInputStream(data));
-          if (fileContents.flags != null) {
-            for (Map.Entry<String, JsonElement> e: fileContents.flags.entrySet()) {
-              builder.add(VersionedDataKind.FEATURES, FlagFactory.flagFromJson(e.getValue()));
-            }
-          }
-          if (fileContents.flagValues != null) {
-            for (Map.Entry<String, JsonElement> e: fileContents.flagValues.entrySet()) {
-              builder.add(VersionedDataKind.FEATURES, FlagFactory.flagWithValue(e.getKey(), e.getValue()));
-            }
-          }
-          if (fileContents.segments != null) {
-            for (Map.Entry<String, JsonElement> e: fileContents.segments.entrySet()) {
-              builder.add(VersionedDataKind.SEGMENTS, FlagFactory.segmentFromJson(e.getValue()));
-            }
-          }
-        } catch (FileDataException e) {
-          throw new FileDataException(e.getMessage(), e.getCause(), p);
-        } catch (IOException e) {
-          throw new FileDataException(null, e, p);
-        }
-      }
-    }
-  }
-  
-  /**
-   * Internal data structure that organizes flag/segment data into the format that the feature store
-   * expects. Will throw an exception if we try to add the same flag or segment key more than once.
-   */
-  static final class DataBuilder {
-    private final Map<VersionedDataKind<?>, Map<String, ? extends VersionedData>> allData = new HashMap<>();
-    
-    public Map<VersionedDataKind<?>, Map<String, ? extends VersionedData>> build() {
-      return allData;
-    }
-    
-    public void add(VersionedDataKind<?> kind, VersionedData item) throws FileDataException {
-      @SuppressWarnings("unchecked")
-      Map<String, VersionedData> items = (Map<String, VersionedData>)allData.get(kind);
-      if (items == null) {
-        items = new HashMap<String, VersionedData>();
-        allData.put(kind, items);
-      }
-      if (items.containsKey(item.getKey())) {
-        throw new FileDataException("in " + kind.getNamespace() + ", key \"" + item.getKey() + "\" was already defined", null, null);
-      }
-      items.put(item.getKey(), item);
     }
   }
 }
