+++ src/main/java/com/launchdarkly/client/files/package-info.java
@@ -1,4 +1,6 @@
 /**
- * Deprecated package replaced by {@link com.launchdarkly.client.integrations.FileData}. 
+ * Package for the file data source component, which may be useful in tests.
+ * <p>
+ * The entry point is {@link com.launchdarkly.client.files.FileComponents}. 
  */
 package com.launchdarkly.client.files;
