+++ src/test/java/com/launchdarkly/client/files/DataLoaderTest.java
@@ -1,4 +1,4 @@
-package com.launchdarkly.client.integrations;
+package com.launchdarkly.client.files;
 
 import com.google.common.collect.ImmutableList;
 import com.google.gson.Gson;
@@ -6,9 +6,6 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.launchdarkly.client.VersionedData;
 import com.launchdarkly.client.VersionedDataKind;
-import com.launchdarkly.client.integrations.FileDataSourceImpl.DataBuilder;
-import com.launchdarkly.client.integrations.FileDataSourceImpl.DataLoader;
-import com.launchdarkly.client.integrations.FileDataSourceParsing.FileDataException;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -17,13 +14,12 @@ import java.util.Map;
 
 import static com.launchdarkly.client.VersionedDataKind.FEATURES;
 import static com.launchdarkly.client.VersionedDataKind.SEGMENTS;
-import static com.launchdarkly.client.integrations.FileDataSourceTestData.FLAG_VALUE_1_KEY;
-import static com.launchdarkly.client.integrations.FileDataSourceTestData.resourceFilePath;
+import static com.launchdarkly.client.files.TestData.FLAG_VALUE_1_KEY;
+import static com.launchdarkly.client.files.TestData.resourceFilePath;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
 
-@SuppressWarnings("javadoc")
 public class DataLoaderTest {
   private static final Gson gson = new Gson();
   private DataBuilder builder = new DataBuilder();
@@ -74,7 +70,7 @@ public class DataLoaderTest {
       DataLoader ds = new DataLoader(ImmutableList.of(resourceFilePath("flag-only.json"),
           resourceFilePath("flag-with-duplicate-key.json")));
       ds.load(builder);
-    } catch (FileDataException e) {
+    } catch (DataLoaderException e) {
       assertThat(e.getMessage(), containsString("key \"flag1\" was already defined"));
     }
   }
@@ -85,7 +81,7 @@ public class DataLoaderTest {
       DataLoader ds = new DataLoader(ImmutableList.of(resourceFilePath("flag-only.json"),
           resourceFilePath("value-with-duplicate-key.json")));
       ds.load(builder);
-    } catch (FileDataException e) {
+    } catch (DataLoaderException e) {
       assertThat(e.getMessage(), containsString("key \"flag1\" was already defined"));
     }
   }
@@ -96,7 +92,7 @@ public class DataLoaderTest {
       DataLoader ds = new DataLoader(ImmutableList.of(resourceFilePath("segment-only.json"),
           resourceFilePath("segment-with-duplicate-key.json")));
       ds.load(builder);
-    } catch (FileDataException e) {
+    } catch (DataLoaderException e) {
       assertThat(e.getMessage(), containsString("key \"seg1\" was already defined"));
     }
   }
