+++ src/test/java/com/launchdarkly/client/files/FileDataSourceTest.java
@@ -1,4 +1,4 @@
-package com.launchdarkly.client.integrations;
+package com.launchdarkly.client.files;
 
 import com.launchdarkly.client.FeatureStore;
 import com.launchdarkly.client.InMemoryFeatureStore;
@@ -14,28 +14,28 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.concurrent.Future;
 
-import static com.launchdarkly.client.integrations.FileDataSourceTestData.ALL_FLAG_KEYS;
-import static com.launchdarkly.client.integrations.FileDataSourceTestData.ALL_SEGMENT_KEYS;
-import static com.launchdarkly.client.integrations.FileDataSourceTestData.getResourceContents;
-import static com.launchdarkly.client.integrations.FileDataSourceTestData.resourceFilePath;
+import static com.launchdarkly.client.files.FileComponents.fileDataSource;
+import static com.launchdarkly.client.files.TestData.ALL_FLAG_KEYS;
+import static com.launchdarkly.client.files.TestData.ALL_SEGMENT_KEYS;
+import static com.launchdarkly.client.files.TestData.getResourceContents;
+import static com.launchdarkly.client.files.TestData.resourceFilePath;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.fail;
 
-@SuppressWarnings("javadoc")
 public class FileDataSourceTest {
   private static final Path badFilePath = Paths.get("no-such-file.json");
   
   private final FeatureStore store = new InMemoryFeatureStore();
   private final LDConfig config = new LDConfig.Builder().build();
-  private final FileDataSourceBuilder factory;
+  private final FileDataSourceFactory factory;
   
   public FileDataSourceTest() throws Exception {
     factory = makeFactoryWithFile(resourceFilePath("all-properties.json"));
   }
   
-  private static FileDataSourceBuilder makeFactoryWithFile(Path path) {
-    return FileData.dataSource().filePaths(path);
+  private static FileDataSourceFactory makeFactoryWithFile(Path path) {
+    return fileDataSource().filePaths(path);
   }
   
   @Test
@@ -94,7 +94,7 @@ public class FileDataSourceTest {
   @Test
   public void modifiedFileIsNotReloadedIfAutoUpdateIsOff() throws Exception {
     File file = makeTempFlagFile();
-    FileDataSourceBuilder factory1 = makeFactoryWithFile(file.toPath());
+    FileDataSourceFactory factory1 = makeFactoryWithFile(file.toPath());
     try {
       setFileContents(file, getResourceContents("flag-only.json"));
       try (UpdateProcessor fp = factory1.createUpdateProcessor("", config, store)) {
@@ -115,7 +115,7 @@ public class FileDataSourceTest {
   @Test
   public void modifiedFileIsReloadedIfAutoUpdateIsOn() throws Exception {
     File file = makeTempFlagFile();
-    FileDataSourceBuilder factory1 = makeFactoryWithFile(file.toPath()).autoUpdate(true);
+    FileDataSourceFactory factory1 = makeFactoryWithFile(file.toPath()).autoUpdate(true);
     long maxMsToWait = 10000;
     try {
       setFileContents(file, getResourceContents("flag-only.json"));  // this file has 1 flag
@@ -142,7 +142,7 @@ public class FileDataSourceTest {
   public void ifFilesAreBadAtStartTimeAutoUpdateCanStillLoadGoodDataLater() throws Exception {
     File file = makeTempFlagFile();
     setFileContents(file, "not valid");
-    FileDataSourceBuilder factory1 = makeFactoryWithFile(file.toPath()).autoUpdate(true);
+    FileDataSourceFactory factory1 = makeFactoryWithFile(file.toPath()).autoUpdate(true);
     long maxMsToWait = 10000;
     try {
       try (UpdateProcessor fp = factory1.createUpdateProcessor("", config, store)) {
