+++ src/test/java/com/launchdarkly/client/files/JsonFlagFileParserTest.java
@@ -1,0 +1,7 @@
+package com.launchdarkly.client.files;
+
+public class JsonFlagFileParserTest extends FlagFileParserTestBase {
+  public JsonFlagFileParserTest() {
+    super(new JsonFlagFileParser(), ".json");
+  }
+}
