+++ src/test/java/com/launchdarkly/client/files/TestData.java
@@ -1,4 +1,4 @@
-package com.launchdarkly.client.integrations;
+package com.launchdarkly.client.files;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
@@ -14,8 +14,7 @@ import java.nio.file.Paths;
 import java.util.Map;
 import java.util.Set;
 
-@SuppressWarnings("javadoc")
-public class FileDataSourceTestData {
+public class TestData {
   private static final Gson gson = new Gson();
   
   // These should match the data in our test files
@@ -41,11 +40,12 @@ public class FileDataSourceTestData {
   public static final Set<String> ALL_SEGMENT_KEYS = ImmutableSet.of(FULL_SEGMENT_1_KEY);
   
   public static Path resourceFilePath(String filename) throws URISyntaxException {
-    URL resource = FileDataSourceTestData.class.getClassLoader().getResource("filesource/" + filename);
+    URL resource = TestData.class.getClassLoader().getResource("filesource/" + filename);
     return Paths.get(resource.toURI());
   }
   
   public static String getResourceContents(String filename) throws Exception {
     return new String(Files.readAllBytes(resourceFilePath(filename)));
   }
+
 }
