+++ src/test/java/com/launchdarkly/client/files/YamlFlagFileParserTest.java
@@ -1,0 +1,7 @@
+package com.launchdarkly.client.files;
+
+public class YamlFlagFileParserTest extends FlagFileParserTestBase {
+  public YamlFlagFileParserTest() {
+    super(new YamlFlagFileParser(), ".yml");
+  }
+}
