--- src/test/java/com/launchdarkly/client/integrations/FlagFileParserJsonTest.java
@@ -1,10 +1,0 @@
-package com.launchdarkly.client.integrations;
-
-import com.launchdarkly.client.integrations.FileDataSourceParsing.JsonFlagFileParser;
-
-@SuppressWarnings("javadoc")
-public class FlagFileParserJsonTest extends FlagFileParserTestBase {
-  public FlagFileParserJsonTest() {
-    super(new JsonFlagFileParser(), ".json");
-  }
-}
