--- src/test/java/com/launchdarkly/client/integrations/FlagFileParserYamlTest.java
@@ -1,10 +1,0 @@
-package com.launchdarkly.client.integrations;
-
-import com.launchdarkly.client.integrations.FileDataSourceParsing.YamlFlagFileParser;
-
-@SuppressWarnings("javadoc")
-public class FlagFileParserYamlTest extends FlagFileParserTestBase {
-  public FlagFileParserYamlTest() {
-    super(new YamlFlagFileParser(), ".yml");
-  }
-}
